update build and publish script
